+++ zkml-jolt-core/src/jolt/lookup_trace.rs
@@ -12,18 +12,9 @@ pub trait LookupTrace {
 impl LookupTrace for ONNXCycle {
     fn to_lookup(&self) -> Option<ONNXLookup> {
         match self.instr.opcode {
-            ONNXOpcode::Add => Some(ONNXLookup::Add(ADD(
-                self.ts1_val() as u64,
-                self.ts2_val() as u64,
-            ))),
-            ONNXOpcode::Sub => Some(ONNXLookup::Sub(SUB(
-                self.ts1_val() as u64,
-                self.ts2_val() as u64,
-            ))),
-            ONNXOpcode::Mul => Some(ONNXLookup::Mul(MUL(
-                self.ts1_val() as u64,
-                self.ts2_val() as u64,
-            ))),
+            ONNXOpcode::Add => Some(ONNXLookup::Add(ADD(self.ts1_val(), self.ts2_val()))),
+            ONNXOpcode::Sub => Some(ONNXLookup::Sub(SUB(self.ts1_val(), self.ts2_val()))),
+            ONNXOpcode::Mul => Some(ONNXLookup::Mul(MUL(self.ts1_val(), self.ts2_val()))),
             _ => None,
         }
     }
